+++ Cargo.toml
@@ -3,21 +3,20 @@ resolver = "2"
 members = [
     "simnode",
     "runtime-api",
-
-    "integration/tests",
+    "tests",
 
     # parachain
-    "integration/parachain/node",
-    "integration/parachain/runtime",
+    "examples/parachain/node",
+    "examples/parachain/runtime",
 
     # aura
-    "integration/aura/node",
-    "integration/aura/runtime",
+    "examples/aura/node",
+    "examples/aura/runtime",
 
      # babe
-    "integration/babe/node",
-    "integration/babe/runtime",
-    "integration/babe/rpc",
+    "examples/babe/node",
+    "examples/babe/runtime",
+    "examples/babe/rpc",
 ]
 
 [workspace.dependencies]
