+++ examples/aura/node/src/command.rs
