+++ examples/aura/node/src/lib.rs
