+++ examples/aura/node/src/main.rs
