+++ examples/babe/rpc/src/lib.rs
