+++ examples/babe/runtime/src/assets_api.rs
