+++ simnode/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "sc-simnode"
-version = "1.6.0"
+version = "1.6.1"
 authors = ["Polytope Labs <hello@polytope.technology>"]
 edition = "2021"
 license = "Apache-2.0"
@@ -30,7 +30,7 @@ sc-client-api = { workspace = true }
 sc-rpc-server = { workspace = true }
 
 # consensus
-sc-consensus-babe = { workspace = true, optional = true }
+sc-consensus-babe = { workspace = true }
 manual-seal = { workspace = true }
 sc-consensus = { workspace = true }
 sc-consensus-slots = { workspace = true }
@@ -39,7 +39,7 @@ sc-consensus-aura = { workspace = true, default-features = true }
 # primitive deps
 sp-consensus-aura = { workspace = true, default-features = true }
 sp-consensus-slots = { workspace = true, default-features = true }
-sp-consensus-babe = { workspace = true, optional = true, default-features = true }
+sp-consensus-babe = { workspace = true, default-features = true }
 sp-core = { workspace = true, default-features = true }
 sp-blockchain = { workspace = true, default-features = true }
 sp-block-builder = { workspace = true, default-features = true }
@@ -95,4 +95,4 @@ std = []
 # consensus
 parachain = []
 aura = []
-babe = ["sp-consensus-babe", "sc-consensus-babe"]
+babe = []
