+++ tests/src/aura.rs
@@ -1,14 +1,13 @@
 #![cfg(test)]
 
-use crate::codegen::parachain::{
+use crate::codegen::aura::{
 	api,
 	api::{
 		balances::events::Transfer,
-		parachain_system::events::{ValidationFunctionApplied, ValidationFunctionStored},
 		runtime_types::{
-			frame_system::pallet::Call, parachain_runtime::RuntimeCall,
-			sp_weights::weight_v2::Weight,
+			aura_runtime::RuntimeCall, frame_system::pallet::Call, sp_weights::weight_v2::Weight,
 		},
+		system::events::CodeUpdated,
 	},
 };
 use anyhow::anyhow;
@@ -96,10 +95,10 @@ async fn runtime_upgrades() -> Result<(), anyhow::Error> {
 	let old_version = client.rpc().runtime_version(None).await?;
 	assert_eq!(old_version.spec_version, 1);
 
-	let code = include_bytes!("../../../assets/parachain-runtime-upgrade.wasm").to_vec();
+	let code = include_bytes!("../../assets/aura-runtime-upgrade.wasm").to_vec();
 
 	let call = client.tx().call_data(&api::tx().sudo().sudo_unchecked_weight(
-		RuntimeCall::System(Call::set_code_without_checks { code }),
+		RuntimeCall::System(Call::set_code { code }),
 		Weight { ref_time: 0, proof_size: 0 },
 	))?;
 
@@ -113,27 +112,16 @@ async fn runtime_upgrades() -> Result<(), anyhow::Error> {
 		.await?;
 	let submittable = SubmittableExtrinsic::from_bytes(client.clone(), extrinsic.0);
 	let events = submittable.submit_and_watch().await?.wait_for_finalized_success().await?;
-	// assert that the event was stored
+	// assert that the event was emitted
 	events
-		.find::<ValidationFunctionStored>()
+		.find::<CodeUpdated>()
 		.collect::<Result<Vec<_>, subxt::Error>>()?
 		.pop()
-		.ok_or_else(|| anyhow!("ValidationFunctionStored event not found!"))?;
-
-	client.rpc().request::<()>("simnode_upgradeSignal", rpc_params![true]).await?;
-
-	let events = client.events().at_latest().await?;
-
-	// assert that the event was stored
-	events
-		.find::<ValidationFunctionApplied>()
-		.collect::<Result<Vec<_>, subxt::Error>>()?
-		.pop()
-		.ok_or_else(|| anyhow!("ValidationFunctionApplied event not found!"))?;
+		.ok_or_else(|| anyhow!("transfer event not found!"))?;
 
 	// assert the version
 	let new_version = client.rpc().runtime_version(None).await?;
-	assert_eq!(new_version.spec_version, 100);
+	assert_eq!(new_version.spec_version, 1000);
 
 	// try to create 10 blocks to assert that the runtime still works.
 	for _ in 0..10 {
