+++ tests/src/codegen/aura.rs
